Porting googlenet
