+++ torchvision/models/alexnet.py
@@ -9,8 +9,8 @@ from ._api import WeightsEnum, Weights
 from ._meta import _IMAGENET_CATEGORIES
 from ._utils import handle_legacy_interface, _ovewrite_named_param
 
-from torchvision.transforms import ImageClassificationEval
-from torchvision.transforms.functional import InterpolationMode
+from ..transforms import ImageClassificationEval, InterpolationMode
+
 
 __all__ = ["AlexNet", "AlexNet_Weights", "alexnet"]
